feat: Criado os métodos nas classes responsáveis para adicionar e deletar uma pessoa

+++ src/main/java/com/trybe/acc/java/programamilhas/dao/PessoaDao.java
@@ -50,7 +50,7 @@ public class PessoaDao {
    * 
    */
   @Transactional
-  public void deletar(Long id) {
+  public void deletar(Integer id) {
     Pessoa pessoa = entityManager.find(Pessoa.class, id);
     entityManager.remove(pessoa);
     return;
